% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/network.R
\name{AddConnectivity}
\alias{AddConnectivity}
\alias{AddConnectivity.CorrelationNetwork}
\alias{AddConnectivity.default}
\title{Calculate the connectivity values of genes.}
\usage{
AddConnectivity(object, ...)

\method{AddConnectivity}{CorrelationNetwork}(object)

\method{AddConnectivity}{default}(object, ...)
}
\arguments{
\item{object}{A CorrelationNetwork object.}

\item{...}{further arguments to be passed to or from other methods.}
}
\value{
A new CorrelationNetwork object.
}
\description{
Calculate the connectivity values of genes.

Calculate the connectivity values for genes.
}
