% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/network.R
\name{AddNetwork}
\alias{AddNetwork}
\alias{AddNetwork.ExpAssayFrame}
\alias{AddNetwork.default}
\title{Construct a correlation network..}
\usage{
AddNetwork(object, ...)

\method{AddNetwork}{ExpAssayFrame}(assay, power)

\method{AddNetwork}{default}(object, ...)
}
\arguments{
\item{object}{An object.}

\item{...}{further arguments to be passed to or from other methods.}

\item{assay}{An ExpAssayFrame object.}

\item{power}{(A length-1 numeric) Soft thresholding power
for the adjacency function in WGCNA analysis.}
}
\value{
A CorrelationNetwork object.
}
\description{
Construct a correlation network..

Construct a correlation network by the gene expression profiles.
}
