% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/preprocessing.R
\name{AddPhenotype}
\alias{AddPhenotype}
\alias{AddPhenotype.ExpAssayTable}
\alias{AddPhenotype.ExpAssayFrame}
\alias{AddPhenotype.default}
\title{Add the phenotype information to assays.}
\usage{
AddPhenotype(object, ...)

\method{AddPhenotype}{ExpAssayTable}(object, phenotype)

\method{AddPhenotype}{ExpAssayFrame}(object, phenotype)

\method{AddPhenotype}{default}(object, ...)
}
\arguments{
\item{object}{An object of class ExpAssayFrame.}

\item{...}{further arguments to be passed to or from other methods.}

\item{phenotype}{An object of class ExperimentInfo.}
}
\value{
A new object of class ExpAssayTable.

A new object of class ExpAssayFrame.
}
\description{
Add the phenotype information to assays.

Add a phenotype table of the samples in the assay.

Add a phenotype table of the samples in the assay.
}
