% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{as_data_frame}
\alias{as_data_frame}
\title{Coerce to a data.frame.}
\usage{
as_data_frame(x, row.names = NULL, ...)
}
\arguments{
\item{x}{An object.}

\item{row.names}{A length-1 character specifying the column to be row names,
or a call, e.g. \code{x$ID} or \code{x[["ID"]]}, specifying the column. Else, this param
will be passed to as.data.frame().}

\item{...}{additional arguments to be passed to as.data.frame().}
}
\value{
A data.frame object.
}
\description{
Coerce to a data.frame, and move a column
to be row names if specifying the column.
}
\examples{
a = data.table::data.table(name = LETTERS[1:10], value = runif(10))
as_data_frame(a, a$name)

}
