% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traits.R
\name{BindModuleSignificance.CorrelationNetwork}
\alias{BindModuleSignificance.CorrelationNetwork}
\title{Combine the module-trait correlation and significance results.}
\usage{
\method{BindModuleSignificance}{CorrelationNetwork}(this, that)
}
\arguments{
\item{this}{A CorrelationNetwork object.}

\item{that}{Another CorrelationNetwork object.}
}
\value{
A new CorrelationNetwork object.
}
\description{
Combine the module-trait correlation and significance results
from two CorrelationNetwork objects.
}
