% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{CorrelationNetwork-class}
\alias{CorrelationNetwork-class}
\title{The CorrelationNetwork class.}
\description{
A list of data frames of experiment data, with
a list of corresponding correlation networks.
}
