% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/plotting.R
\name{DistributionPlot}
\alias{DistributionPlot}
\alias{DistributionPlot.ExpAssayFrame}
\alias{DistributionPlot.default}
\title{Draw the distribution.}
\usage{
DistributionPlot(object, ...)

\method{DistributionPlot}{ExpAssayFrame}(assay, samples, file)

\method{DistributionPlot}{default}(object, ...)
}
\arguments{
\item{object}{An object.}

\item{...}{further arguments to be passed to or from other methods.}
}
\description{
Draw the distribution.
}
