% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traits.R
\name{GeneSignificance.CorrelationNetwork}
\alias{GeneSignificance.CorrelationNetwork}
\title{Compute gene-trait correlation and significance.}
\usage{
\method{GeneSignificance}{CorrelationNetwork}(object, samples, traits, prefix)
}
\arguments{
\item{object}{A CorrelationNetwork object.}

\item{samples}{A character vector specifying the rows for analysis.}

\item{traits}{A character vector specifying the columns of traits for analysis.}

\item{prefix}{(A length-1 character) A prefix representing the disease.}
}
\value{
A new CorrelationNetwork object.
}
\description{
Compute gene-trait correlation and significance.
}
