% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_geneinfo}
\alias{get_geneinfo}
\title{Get the gene information from org.Hs.eg.db.}
\usage{
get_geneinfo(d.t, remove_dup = TRUE)
}
\arguments{
\item{d.t}{A data.table.}

\item{remove_dup}{(A length-1 logical) Remove duplicate genes or not.}
}
\value{
A data.table.
}
\description{
Get the gene information from org.Hs.eg.db.
}
\examples{
\dontrun{
d.t = data.table::data.table(gene = "ATP5A1")
d.t = get_geneinfo(d.t)
}

}
