% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/network.R
\name{GetConnectivity}
\alias{GetConnectivity}
\alias{GetConnectivity.CorrelationNetwork}
\alias{GetConnectivity.default}
\title{Return the degree of each node.}
\usage{
GetConnectivity(object, ...)

\method{GetConnectivity}{CorrelationNetwork}(object, geneinfo = TRUE, index = 1)

\method{GetConnectivity}{default}(object, ...)
}
\arguments{
\item{object}{A CorrelationNetwork object.}

\item{...}{further arguments to be passed to or from other methods.}

\item{geneinfo}{(A length-1 logical) Add gene annotation or not.}

\item{index}{A length-1 numeric or character vector specifying the frame. (default: 1)}
}
\value{
A new CorrelationNetwork object.
}
\description{
Return both the whole network and
the intramodular connectivity for all genes.
}
