% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/network.R
\name{GetHubGenes}
\alias{GetHubGenes}
\alias{GetHubGenes.CorrelationNetwork}
\alias{GetHubGenes.default}
\title{Find hub genes.}
\usage{
GetHubGenes(object, ...)

\method{GetHubGenes}{CorrelationNetwork}(
  object,
  intramodular.ratio.threshold = 0.9,
  ratio.threshold = 0,
  top.n,
  index = 1,
  geneinfo = TRUE
)

\method{GetHubGenes}{default}(object, ...)
}
\arguments{
\item{object}{A CorrelationNetwork object.}

\item{...}{further arguments to be passed to or from other methods.}

\item{intramodular.ratio.threshold}{Choose the genes with the connectivity higher
than 0.9 * top_connectivity_within_the_module,
when intramodular.ratio.threshold = 0.9.}

\item{ratio.threshold}{Choose the genes with the connectivity higher
than 0.9 * top_connectivity_across_all_modules,
when ratio.threshold = 0.9.}

\item{top.n}{(A length-1 integer) Choose n top genes in each modules.}

\item{index}{A length-1 numeric or character vector specifying the frame. (default: 1)}

\item{geneinfo}{(A length-1 logical) Add gene annotation or not.}
}
\value{
A data.table with hub genes.
}
\description{
Find hub genes.

Find hub genes in each module of a correlation network.
}
