% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/network.R
\name{GetNetwork}
\alias{GetNetwork}
\alias{GetNetwork.ExpAssayFrame}
\alias{GetNetwork.default}
\title{Construct a correlation network..}
\usage{
GetNetwork(object, ...)

\method{GetNetwork}{ExpAssayFrame}(assay, power)

\method{GetNetwork}{default}(object, ...)
}
\arguments{
\item{object}{An object.}

\item{...}{further arguments to be passed to or from other methods.}
}
\description{
Construct a correlation network..
}
