% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traits.R
\name{GetRelatedHubGenes.CorrelationNetwork}
\alias{GetRelatedHubGenes.CorrelationNetwork}
\title{Get hub genes in the modules associated with traits of interest.}
\usage{
\method{GetRelatedHubGenes}{CorrelationNetwork}(object, traits, prefix, index = 1, ...)
}
\arguments{
\item{object}{A CorrelationNetwork object.}

\item{traits}{A character vector specifying the columns of traits for analysis.}

\item{prefix}{(A length-1 character) A prefix representing the disease.}

\item{index}{A length-1 numeric or character vector specifying the frame. (default: 1)}

\item{...}{further arguments to be passed to \code{GetHubGenes()}.}
}
\value{
A data.table with hub genes.
}
\description{
Get hub genes in the modules associated with traits of interest.
}
