% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/traits.R
\name{GetRelatedHubGenes}
\alias{GetRelatedHubGenes}
\alias{GetRelatedHubGenes.default}
\title{Get hub genes in the modules associated with traits of interest.}
\usage{
GetRelatedHubGenes(object, ...)

\method{GetRelatedHubGenes}{default}(object, ...)
}
\arguments{
\item{object}{An object.}

\item{...}{further arguments to be passed to or from other methods.}
}
\description{
Get hub genes in the modules associated with traits of interest.
}
