% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/traits.R
\name{GetSignificantGenes}
\alias{GetSignificantGenes}
\alias{GetSignificantGenes.CorrelationNetwork}
\alias{GetSignificantGenes.default}
\title{Get genes associated with traits of interest.}
\usage{
GetSignificantGenes(object, ...)

\method{GetSignificantGenes}{CorrelationNetwork}(object, traits, prefix, index = 1)

\method{GetSignificantGenes}{default}(object, ...)
}
\arguments{
\item{object}{A CorrelationNetwork object.}

\item{...}{further arguments to be passed to or from other methods.}

\item{traits}{A character vector specifying the columns of traits for analysis.}

\item{prefix}{(A length-1 character) A prefix representing the disease.}

\item{index}{A length-1 numeric or character vector specifying the frame. (default: 1)}
}
\value{
A data.table with genes.
}
\description{
Get genes associated with traits of interest.

Get the genes associated with traits of interest.
}
