% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/plotting.R
\name{Histogram}
\alias{Histogram}
\alias{Histogram.ExpAssayFrame}
\alias{Histogram.default}
\title{Plot a distribution.}
\usage{
Histogram(object, ...)

\method{Histogram}{ExpAssayFrame}(
  object,
  samples,
  genes,
  index = 1,
  preview = FALSE,
  title,
  file,
  plot.width = NA,
  plot.height = NA
)

\method{Histogram}{default}(object, ...)
}
\arguments{
\item{object}{An ExpAssayFrame object.}

\item{...}{further arguments to be passed to or from other methods.}

\item{samples}{A character vector specifying the samples of interest.}

\item{genes}{A character vector specifying the genes of interest.}

\item{index}{A length-1 numeric or character vector specifying the frame. (default: 1)}

\item{preview}{(A length-1 logical) If TRUE, return a ggplot object; else
if FALSE, save the plot to a file. (default: FALSE)}

\item{title}{(A length-1 character) Title for the plot.}

\item{file}{(A length-1 character) File path to save the plot.}

\item{plot.width}{(A length-1 numeric) Set the plot width.
If not supplied, use the size of current graphics device.}

\item{plot.height}{(A length-1 numeric) Set the plot height.
If not supplied, use the size of current graphics device.}
}
\value{
Path to the output file, or a ggplot object when preview=TRUE.
}
\description{
Plot a distribution.

Plot the distribution of expression levels in the assay.
}
