% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{is_missing_in_ms}
\alias{is_missing_in_ms}
\title{Check missing values in mass spectrometry data.}
\usage{
is_missing_in_ms(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A logical vector.
}
\description{
Check if the values are missing in MS data
(NA in labelled MS data, and 0 in label-free MS data).
}
\examples{
\dontrun{
is_missing_in_ms(c(1, 0, NA, -Inf))
}

}
