% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{is_missing_value}
\alias{is_missing_value}
\title{Check missing values.}
\usage{
is_missing_value(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A logical vector.
}
\description{
Check if the values are missing
(NA in labelled MS data, and 0 in label-free MS data).
}
\examples{
\dontrun{
is_missing_value(c(1, 0, 0, 1))
}

}
