% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/preprocessing.R
\name{LogNorm}
\alias{LogNorm}
\alias{LogNorm.ExpAssayFrame}
\alias{LogNorm.default}
\title{Perform log normalization.}
\usage{
LogNorm(object, ...)

\method{LogNorm}{ExpAssayFrame}(object, inverse = TRUE)

\method{LogNorm}{default}(object, ...)
}
\arguments{
\item{object}{An object of class ExpAssayFrame.}

\item{...}{further arguments to be passed to or from other methods.}

\item{inverse}{(logical) Compute log2 (FALSE) or -log2 (TRUE).}
}
\value{
A new object of class ExpAssayFrame.
}
\description{
Perform log normalization.

Perform log normalization of an expression profile.
}
\examples{
\dontrun{
new.assay = LogNorm(old.assay)
}

}
