% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/preprocessing.R
\name{LogTransform}
\alias{LogTransform}
\alias{LogTransform.ExpAssayFrame}
\alias{LogTransform.default}
\title{Perform log transformation.}
\usage{
LogTransform(object, ...)

\method{LogTransform}{ExpAssayFrame}(object, inverse = FALSE)

\method{LogTransform}{default}(object, ...)
}
\arguments{
\item{object}{An object of class ExpAssayFrame.}

\item{...}{further arguments to be passed to or from other methods.}

\item{inverse}{(logical) Compute log2 (FALSE) or -log2 (TRUE).}
}
\value{
A new object of class ExpAssayFrame.
}
\description{
Perform log transformation.

Perform log2 transformation of an expression profile.
}
\examples{
\dontrun{
new.assay = LogTransform(old.assay)
}

}
