% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traits.R
\name{ModuleMembership.CorrelationNetwork}
\alias{ModuleMembership.CorrelationNetwork}
\title{Compute gene-module correlation and significance.}
\usage{
\method{ModuleMembership}{CorrelationNetwork}(object, samples, prefix)
}
\arguments{
\item{object}{A CorrelationNetwork object.}

\item{samples}{A character vector specifying the rows for analysis.}

\item{prefix}{(A length-1 character) A prefix representing the disease.}
}
\value{
A new CorrelationNetwork object.
}
\description{
Compute gene-module correlation and significance.
}
