% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{ModuleTraitHeatmap}
\alias{ModuleTraitHeatmap}
\title{Plot a module-trait heatmap.}
\usage{
ModuleTraitHeatmap(object, ...)
}
\arguments{
\item{object}{An object.}

\item{...}{further arguments to be passed to or from other methods.}
}
\description{
Plot a module-trait heatmap.
}
