% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/preprocessing.R
\name{Normalize}
\alias{Normalize}
\alias{Normalize.ExpAssayFrame}
\alias{Normalize.default}
\title{Perform normalization.}
\usage{
Normalize(object, ...)

\method{Normalize}{ExpAssayFrame}(object, method = "center.median")

\method{Normalize}{default}(object, ...)
}
\arguments{
\item{object}{An object of class ExpAssayFrame.}

\item{...}{further arguments to be passed to or from other methods.}

\item{method}{Choose the method to use.}
}
\value{
A new object of class ExpAssayFrame.
}
\description{
Perform normalization.

Normalize a log-transformed expression profile.
}
\examples{
\dontrun{
new.assay = Normalize(LogTransform(old.assay), method = "center.median")
}

}
