% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/preprocessing.R
\name{QC}
\alias{QC}
\alias{QC.ProteinGroups}
\alias{QC.default}
\title{Perform quality control for assays.}
\usage{
QC(object, ...)

\method{QC}{ProteinGroups}(object, min.unique.peptides = 2, min.fraction = 0.5)

\method{QC}{default}(object, ...)
}
\arguments{
\item{object}{An object of class ProteinGroups.}

\item{...}{further arguments to be passed to or from other methods.}

\item{min.unique.peptides}{A threshold for the number of unique peptides.}

\item{min.fraction}{A minimum fraction of non-missing samples for a gene
to be considered good.}
}
\value{
A new object of class ProteinGroups.
}
\description{
Perform quality control for assays.

Quality control of MS-based proteomics data.
}
\examples{
\dontrun{
new.Assay = QC(old.Assay)
}

}
