% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{ReadExperimentalDesign}
\alias{ReadExperimentalDesign}
\title{Read the experiment information of mass spectrometry data.}
\usage{
ReadExperimentalDesign(data.dir = ".")
}
\arguments{
\item{data.dir}{(length-1 character vector)
Directory containing the MaxQuantOutput data.}
}
\value{
(An ExperimentInfo object)
A list containing the experiment information.
}
\description{
Load the data from the experimentalDesign.txt or summary.txt file in
MaxQuantOutput/ and create an ExperimentInfo object.
}
\examples{
\dontrun{
object <- ReadExperimentalDesign('.')
}

}
