% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reader.R
\name{ReadPhenotypeTable}
\alias{ReadPhenotypeTable}
\title{Read the phenotype information of the samples.}
\usage{
ReadPhenotypeTable(file)
}
\arguments{
\item{file}{A length-1 character specifying the file path.}
}
\value{
An ExperimentInfo object.
}
\description{
Read the phenotype information of the samples.
}
