% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{ReadProteinGroups}
\alias{ReadProteinGroups}
\title{Read the protein-level mass spectrometry data.}
\usage{
ReadProteinGroups(
  data.dir = ".",
  column.prefix = c("Ratio H/L normalized", "LFQ intensity")
)
}
\arguments{
\item{data.dir}{Directory containing the MaxQuantOutput data.}

\item{column.prefix}{A character vector with the prefixes of the expression columns to extract.}
}
\value{
A list of data.tables containing the expression data.
}
\description{
Load the data from the proteinGroups.txt file in MaxQuantOutput/ and create a ProteinGroups object.
}
\examples{
\dontrun{
object <- ReadProteinGroups('.', col = 'Ratio H/L normalized')
}

}
