% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/plotting.R
\name{SampleTree}
\alias{SampleTree}
\alias{SampleTree.ExpAssayFrame}
\alias{SampleTree.default}
\title{Perform sample clustering.}
\usage{
SampleTree(object, ...)

\method{SampleTree}{ExpAssayFrame}(
  assay,
  file,
  title = "Sample Clustering",
  hline = 50,
  plot.width = 20,
  plot.height = 12
)

\method{SampleTree}{default}(object, ...)
}
\arguments{
\item{object}{An object.}

\item{...}{further arguments to be passed to or from other methods.}

\item{assay}{An ExpAssayFrame object.}

\item{file}{(A length-1 character vector) File path to save the plot.}

\item{title}{(A length-1 character vector) Title for the plot.}

\item{hline}{(A length-1 numeric) Set the position of a red horizontal line.}

\item{plot.width}{(A length-1 numeric) Set the plot width.}

\item{plot.height}{(A length-1 numeric) Set the plot height.}
}
\value{
Path to the output file.
}
\description{
Perform sample clustering.

Make a plot of sample clustering using an expression profile.
}
