% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{Show}
\alias{Show}
\title{Show the variable name and the value.}
\usage{
Show(x)
}
\arguments{
\item{x}{The object to display.}
}
\value{
The input.
}
\description{
Show the variable name and the value.
}
\examples{
data.dir = '/path/to/dir'
Show(data.dir)

}
