% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/preprocessing.R
\name{Subset}
\alias{Subset}
\alias{Subset.ExpAssayTable}
\alias{Subset.default}
\title{Subset assays.}
\usage{
Subset(object, ...)

\method{Subset}{ExpAssayTable}(object, samples)

\method{Subset}{default}(object, ...)
}
\arguments{
\item{object}{An object of class ExpAssayTable.}

\item{...}{further arguments to be passed to or from other methods.}

\item{samples}{(character) The samples for subsetting.}
}
\value{
A new object of class ExpAssayTable.
}
\description{
Subset assays.

Subsetting an ExpAssayTable object.
}
\examples{
\dontrun{
new.Assay = Subset(old.Assay, samples = c("sample_1", "sample_2", "sample_3"))
}

}
