% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targetIDbyWGCNA-package.R
\docType{package}
\name{targetIDbyWGCNA}
\alias{targetIDbyWGCNA}
\alias{_PACKAGE}
\alias{targetIDbyWGCNA-package}
\title{Target Identification Using the WGCNA Method}
\description{
Target identification is an essential first step in drug
discovery. This package implements convenient functions for performing
target identification tasks on gene expression data using the WGCNA method.
}
\author{
\strong{Maintainer}: Chen Liang \email{chen.liang@ghddi.org}

}
\keyword{internal}
