% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/network.R
\name{AddConnectivity}
\alias{AddConnectivity}
\alias{AddConnectivity.CorrelationNetwork}
\alias{AddConnectivity.default}
\title{Calculate the degree of each node.}
\usage{
AddConnectivity(object, ...)

\method{AddConnectivity}{CorrelationNetwork}(object, geneinfo = FALSE)

\method{AddConnectivity}{default}(object, ...)
}
\arguments{
\item{object}{A CorrelationNetwork object.}

\item{...}{further arguments to be passed to or from other methods.}

\item{geneinfo}{(A length-1 logical) Add gene annotation or not.}
}
\value{
A new CorrelationNetwork object.
}
\description{
Calculate both the whole network and
the intramodular connectivity for genes.
}
