% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{assert_length_1}
\alias{assert_length_1}
\title{Assert a length-1 vector.}
\usage{
assert_length_1(x)
}
\arguments{
\item{x}{The object for checking.}
}
\value{
The first element of the input.
}
\description{
Execute an error if the length of the input is 0 and
execute a warning if the length of the input is bigger
than 1. Return the first element of the input.
}
\examples{
\dontrun{
data.dir = '.'
Show(data.dir)
assert_length_1(data.dir)
data.dir = c('.', '/path/to/dir')
Show(data.dir)
assert_length_1(data.dir)
}

}
