% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/traits.R
\name{BindModuleSignificance}
\alias{BindModuleSignificance}
\alias{BindModuleSignificance.default}
\title{Combine module-trait correlation and significance.}
\usage{
BindModuleSignificance(object, ...)

\method{BindModuleSignificance}{default}(object, ...)
}
\arguments{
\item{object}{An object.}

\item{...}{further arguments to be passed to or from other methods.}
}
\description{
Combine module-trait correlation and significance.
}
