% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\docType{class}
\name{ExpAssayFrame-class}
\alias{ExpAssayFrame-class}
\title{The ExpAssayFrame class.}
\description{
A list of matrices of experiment data, whose rows correspond
to samples and columns to genes.
}
