% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/traits.R
\name{GeneSignificance}
\alias{GeneSignificance}
\alias{GeneSignificance.default}
\title{Compute gene-trait correlation and significance.}
\usage{
GeneSignificance(object, ...)

\method{GeneSignificance}{default}(object, ...)
}
\arguments{
\item{object}{An object.}

\item{...}{further arguments to be passed to or from other methods.}
}
\description{
Compute gene-trait correlation and significance.
}
