% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{ModuleMembership}
\alias{ModuleMembership}
\title{Compute module membership of genes.}
\usage{
ModuleMembership(object, ...)
}
\arguments{
\item{object}{An object.}

\item{...}{further arguments to be passed to or from other methods.}
}
\description{
Compute module membership of genes.
}
