% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/plotting.R
\name{ModulePlot}
\alias{ModulePlot}
\alias{ModulePlot.ExpAssayFrame}
\alias{ModulePlot.default}
\title{Plot a gene tree..}
\usage{
ModulePlot(object, ...)

\method{ModulePlot}{ExpAssayFrame}(assay, file.prefix, plot.width = 12.5, plot.height = 10)

\method{ModulePlot}{default}(object, ...)
}
\arguments{
\item{object}{An object.}

\item{...}{further arguments to be passed to or from other methods.}
}
\description{
Plot a gene tree..

Plot the gene tree.
}
