% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/plotting.R
\name{ModulePlot}
\alias{ModulePlot}
\alias{GeneTree}
\alias{ModulePlot.CorrelationNetwork}
\alias{ModulePlot.default}
\title{Plot a gene tree..}
\usage{
ModulePlot(object, ...)

\method{ModulePlot}{CorrelationNetwork}(assay, index = 1, file.prefix, plot.width = 12.5, plot.height = 10)

\method{ModulePlot}{default}(object, ...)
}
\arguments{
\item{object}{An object.}

\item{...}{further arguments to be passed to or from other methods.}

\item{assay}{A CorrelationNetwork object.}

\item{index}{A length-1 numeric or character vector specifying the frame.}

\item{file.prefix}{(A length-1 character) A prefix for the name of output file.}

\item{plot.width}{(A length-1 numeric) Set the plot width.}

\item{plot.height}{(A length-1 numeric) Set the plot height.}
}
\value{
Path to the output file.
}
\description{
Plot a gene tree..

Plot the gene tree of a gene co-expression network.
}
