% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/traits.R
\name{ModuleSignificance}
\alias{ModuleSignificance}
\alias{ModuleSignificance.default}
\title{Compute module-trait correlation and significance.}
\usage{
ModuleSignificance(object, ...)

\method{ModuleSignificance}{default}(object, ...)
}
\arguments{
\item{object}{An object.}

\item{...}{further arguments to be passed to or from other methods.}
}
\description{
Compute module-trait correlation and significance.
}
