% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/plotting.R
\name{ModuleTraitHeatmap}
\alias{ModuleTraitHeatmap}
\alias{ModuleTraitHeatmap.CorrelationNetwork}
\alias{ModuleTraitHeatmap.default}
\title{Plot a module-trait heatmap.}
\usage{
ModuleTraitHeatmap(object, ...)

\method{ModuleTraitHeatmap}{CorrelationNetwork}(
  object,
  index = 1,
  preview = FALSE,
  title = "Module-Trait Relationships",
  file.prefix,
  plot.width = 8,
  plot.height = 8,
  font.size = 1.5
)

\method{ModuleTraitHeatmap}{default}(object, ...)
}
\arguments{
\item{object}{A CorrelationNetwork object.}

\item{...}{further arguments to be passed to or from other methods.}

\item{index}{A length-1 numeric or character vector specifying the frame. (default: 1)}

\item{preview}{(A length-1 logical) If TRUE, display a plot; else
if FALSE, save the plot to a file. (default: FALSE)}

\item{title}{(A length-1 character vector) Title for the plot.}

\item{file.prefix}{(A length-1 character) A prefix for the name of output file.}

\item{plot.width}{(A length-1 numeric) Set the plot width.}

\item{plot.height}{(A length-1 numeric) Set the plot height.}

\item{font.size}{(A length-1 numeric) Set the font size.}
}
\value{
Path to the output file, or NULL when preview=TRUE.
}
\description{
Plot a module-trait heatmap.

Plot a module-trait heatmap.
}
