% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/plotting.R
\name{SampleTree}
\alias{SampleTree}
\alias{SampleTree.ExpAssayFrame}
\alias{SampleTree.default}
\title{Perform sample clustering.}
\usage{
SampleTree(object, ...)

\method{SampleTree}{ExpAssayFrame}(
  assay,
  file,
  title = "Sample Clustering",
  hline = 50,
  plot.width = 20,
  plot.height = 12
)

\method{SampleTree}{default}(object, ...)
}
\arguments{
\item{object}{An object.}

\item{...}{further arguments to be passed to or from other methods.}

\item{assay}{An ExpAssayFrame object.}

\item{file}{A length-1 character naming the file to draw the plot.}
}
\description{
Perform sample clustering.

Make a plot of sample clustering using an expression profile.
}
