% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{Show}
\alias{Show}
\title{Print the name and the value of a variable.}
\usage{
Show(x)
}
\arguments{
\item{x}{The object to display.}
}
\value{
The input variable.
}
\description{
Print the name and the value of a variable.
}
\examples{
\dontrun{
data.dir = '/path/to/dir'
Show(data.dir)
}

}
