% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/preprocessing.R
\name{Subset}
\alias{Subset}
\alias{Subset.ExperimentAssay}
\title{Subset assays.}
\usage{
Subset(object, ...)

\method{Subset}{ExperimentAssay}(object, samples)
}
\arguments{
\item{object}{An object of class ExperimentAssay.}

\item{...}{further arguments to be passed to or from other methods.}

\item{samples}{(A character vector) The samples for subsetting.}
}
\value{
A new object of class ExperimentAssay.
}
\description{
Subset assays.

Subsetting an ExperimentAssay object.
}
\examples{
\dontrun{
new.Assay = Subset(old.Assay, samples = c("sample_1", "sample_2", "sample_3"))
}

}
