% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targetIDbyWGCNA-package.R
\docType{package}
\name{targetIDbyWGCNA-package}
\alias{targetIDbyWGCNA}
\alias{targetIDbyWGCNA-package}
\title{Target Identification Using the WGCNA Method}
\description{
Target identification is an essential first step in drug
discovery. This package implements convenient functions for performing
target identification tasks on gene expression data using the WGCNA method.
}
\details{
Functions:
\itemize{
\item \code{ReadExperimentalDesign()}
\item \code{ReadProteinGroups()}
}

Classes:
\itemize{
\item \code{ExperimentList-class}
\itemize{
\item \code{ExperimentInfo-class}
\item \code{ExpAssayTable-class}
\itemize{
\item \code{ProteinGroups-class}
}
\item \code{ExpAssayFrame-class}
\itemize{
\item \code{CorrelationNetwork-class}
}
}
}

S3 methods:
\itemize{
\item \code{Tidy()}
\item \code{Subset()}
\item \code{QC()}
\item \code{Reshape()}
\item \code{LogTransform()}
\item \code{Normalize()}
\item \code{Histogram()}
\item \code{SampleTree()}
\item \code{PickThreshold()}
\item \code{AddNetwork()}
\item \code{ModulePlot()}
\item \code{AddConnectivity()}
\item \code{GetHubGenes()}
}
}
\author{
\strong{Maintainer}: Chen Liang \email{chen.liang@ghddi.org}

}
\keyword{internal}
